requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
babel==2.16.0
blinker==1.8.2
boto3==1.35.58
botocore==1.35.58
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-environ==0.11.2
environ==1.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
gast==0.6.0
ghp-import==2.1.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.43
mkdocs-material-extensions==1.3.1
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
paginate==0.5.7
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.6
protobuf==4.25.4
Pygments==2.18.0
pymdown-extensions==10.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
s3transfer==0.10.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==6.0.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
pyspark
awsglue
sklearn
joblib